fix: save psid as bigint in postgres

+++ src/migrations/postgres/up.js
@@ -11,14 +11,14 @@ async function migrateDB() {
 	const 
 		createUsersTableQuery = 
             `CREATE TABLE IF NOT EXISTS users (
-                psid integer NOT NULL UNIQUE,
+                psid BIGINT NOT NULL UNIQUE,
                 name varchar(255) NOT NULL,
                 PRIMARY KEY(psid)
             )`,
 		createMessagesTableQuery = 
             `CREATE TABLE IF NOT EXISTS messages (
                 id serial PRIMARY KEY,
-                user_psid integer NOT NULL,
+                user_psid BIGINT NOT NULL,
                 message_text TEXT NOT NULL,
                 created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
                 FOREIGN KEY (user_psid) REFERENCES users(psid) ON DELETE CASCADE ON UPDATE CASCADE
